Upaded styles,hb menu and layout for the website. (#9)
* Update_style_4_01_2024.css

Update style to make for a more visually pleasing, user-friendly website. Further updates have been made to make the website more responsive and mobile-friendly.

* Update_index_4_01_2024.html

Updated index index.html for a better layout of the elements. Further edits have been made to the contents of the index.html

* Update_musicproduction_4_01_2024.html

Music production.html has been updated for a better layout.

* Addition of the hbmenu.js

Addition of the hbmenu.js for the hamburger menu for smaller screens to work.